Adds 'format' field to the 'show' and 'edit' views.
For now it's just a string, but later may come from a controlled vocabulary.
Making this change requires overwriting views from curation_concerns, and
modifying the presenter for the videos controller. Here I chose to create the
VideoShowPresenter and use that, instead of the default GenericWorkShowPresenter
that lives in the curation_concenrs gem.